% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{compare_heal_list}
\alias{compare_heal_list}
\title{Find the difference in copy number between shared samples of two heal lists.}
\usage{
compare_heal_list(heal_list_1, heal_list_2, n_threads, abs = FALSE)
}
\arguments{
\item{heal_list_1}{List in heal format (such as output from count_heal_data()).}

\item{heal_list_2}{List in heal format (such as output from count_heal_data()).}

\item{n_threads}{Number of threads to use ('1' by default). It parallelize for each shared sample.
The maximum number used is therefore the number of shared samples.}

\item{abs}{Return the absolute value of the differences ('FALSE' by default).}
}
\value{
A heal list with one data table per subgenome holding the difference between shared samples (heal_list_1 - heal_list_2).
}
\description{
Find the difference in copy number between shared samples of two heal lists.
}
