% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_based_correction.R
\name{correct_cn_with_density}
\alias{correct_cn_with_density}
\title{Re-evaluate the copy number at discordant anchors
Goes over all discordant anchor sets and verifies if some of their overlapping bins can be changed based on density. If a change leads to a concordant combination of copy number for the set of progenitors then it is accepted.}
\usage{
correct_cn_with_density(
  densities,
  alignment,
  heal_list,
  n_threads = 1,
  prog_ploidy = 2
)
}
\arguments{
\item{densities}{A list of densities for each polyploid sample (output from get_concordant_density()).}

\item{alignment}{A heal alignment object created with get_heal_alignment().}

\item{heal_list}{Output list in heal format (such as output from count_heal_data()).}

\item{n_threads}{Number of threads to use ('1' by default).}

\item{prog_ploidy}{Ploidy of the progenitors (Assumed to be equal. '2' by default).}
}
\value{
A heal alignment object with the re-evaluated copy number based on empirical densities. The status column indicates what correction procedure (if any) was applied
}
\description{
Re-evaluate the copy number at discordant anchors
Goes over all discordant anchor sets and verifies if some of their overlapping bins can be changed based on density. If a change leads to a concordant combination of copy number for the set of progenitors then it is accepted.
}
