% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_and_correct.R
\name{correct_gc}
\alias{correct_gc}
\title{Correct GC content using loess method}
\usage{
correct_gc(
  heal_list,
  n_threads = 1,
  n_windows = 10,
  loess_span = 0.75,
  prog_ploidy = 2,
  local_normalize = FALSE,
  output_dir = FALSE,
  ymax = 4,
  cex = 1,
  pch = 1,
  alpha = 0.3,
  linewidth = 3,
  device = "pdf"
)
}
\arguments{
\item{heal_list}{List in heal format (such as output from count_heal_data()).}

\item{n_threads}{Number of threads to use ('1' by default).}

\item{n_windows}{Number of gc windows in which to compute the medians ('10' by default).}

\item{loess_span}{The span variable to pass to the loess() function ('0.75' by default).}

\item{prog_ploidy}{of the progenitors (Assumed to be equal. '2' by default).}

\item{local_normalize}{Use local medians to normalize ('FALSE' by default).}

\item{output_dir}{Either the name of a directory to write all plots to (will create one if nonexistent) or FALSE (plots are printed only). Defaults to FALSE.}

\item{ymax}{The y limit of the plots (default = '4').}

\item{cex}{Size of points on plots (default = '1').}

\item{pch}{Point type (default = '1').}

\item{alpha}{Point transparency (default = '0.3')}

\item{linewidth}{Width of loess regression line (default = '3').}

\item{device}{Plot device (as argument to ggplot2::ggsave) ('pdf' by default).}
}
\value{
A heal list with the counts in the 'bins' data tables corrected based on GC content effect.
Also returns a list of lots of corrected and uncorrected counts in the 'gc_correction_plots' list item.
}
\description{
Correct GC content using loess method
}
