% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_based_correction.R
\name{get_concordant_density}
\alias{get_concordant_density}
\title{Get densities from concordant bins.}
\usage{
get_concordant_density(
  alignment,
  heal_list,
  n_threads = 1,
  prog_ploidy = 2,
  n_points = 1000
)
}
\arguments{
\item{alignment}{A heal alignment object created with get_heal_alignment().}

\item{heal_list}{Output list in heal format (such as output from count_heal_data()).}

\item{n_threads}{Number of threads to use ('1' by default).}

\item{prog_ploidy}{Ploidy of the progenitors (Assumed to be equal. '2' by default).}

\item{n_points}{The n parameter for MASS::kde2d(). "Number of grid points in each direction. Can be scalar or a length-2 integer vector". Default is 1000.}
}
\value{
A list with one element per polyploid sample. Each sample has a list with one density distribution for each copy number from 0 to the ploidy of the progenitors times the number of progenitors.
}
\description{
Get the two-dimensional (counts and GC content) kernel density estimation per copy number using bins involved in a uniquely overlapping and concordant anchor
}
