% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_bins.R
\name{get_conserved_anchors}
\alias{get_conserved_anchors}
\title{get_conserved_hits}
\usage{
get_conserved_anchors(genespace_dir)
}
\arguments{
\item{genespace_dir}{Path to a directory containing the syntenicHits output directory for GENESPACE.}
}
\value{
A data table with one row for each anchor set present in all progenitors.
}
\description{
get_conserved_hits
}
