% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_bins.R
\name{get_heal_alignment}
\alias{get_heal_alignment}
\title{Create heal alignment object.}
\usage{
get_heal_alignment(heal_list, genespace_dir, n_threads, prog_ploidy = 2)
}
\arguments{
\item{heal_list}{Output list in heal format (such as output from count_heal_data()).}

\item{genespace_dir}{Path to a directory containing the syntenicHits output directory for GENESPACE.}

\item{n_threads}{Number of threads to use ('1' by default).}

\item{prog_ploidy}{Ploidy of the progenitors (Assumed to be equal. '2' by default).}
}
\value{
A data table with one row for each anchor set present in all progenitors. The bins overlapping each anchor are given along with a consensus copy number at the anchor based on these bins.
}
\description{
Create heal alignment object.
}
