% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sample_stats}
\alias{get_sample_stats}
\title{Get general statistics about the each sample.}
\usage{
get_sample_stats(
  heal_list,
  n_threads = 1,
  method = "median",
  sample_type = FALSE
)
}
\arguments{
\item{heal_list}{Output list in heal format (such as output from count_heal_data()).}

\item{n_threads}{Number of threads to use ('1' by default).}

\item{method}{The type of statistic computed ('median' by default). Options are 'median', 'mean', 'sd', 'is.na' and 'length'.}

\item{sample_type}{Logical on choice to return information about wether the sample is polyploid or not.}
}
\value{
Information about the samples.
}
\description{
Get general statistics about the each sample.
}
