% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_bins.R
\name{map_bins_to_anchors}
\alias{map_bins_to_anchors}
\title{Get the bins overlapping with each anchor gene.}
\usage{
map_bins_to_anchors(heal_list, genespace_dir, n_threads = 1)
}
\arguments{
\item{heal_list}{Output list in heal format (such as output from count_heal_data()).}

\item{genespace_dir}{Path to a directory containing the syntenicHits output directory for GENESPACE.}

\item{n_threads}{Number of threads to use ('1' by default).}
}
\value{
A list containing one data table per progenitor with anchors and overlapping bins.
}
\description{
Get the bins overlapping with each anchor gene.
}
