% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_heal_data.R
\name{parse_counts_bins}
\alias{parse_counts_bins}
\title{Parse feature count output for bins and merge to gc and mappability.}
\usage{
parse_counts_bins(feature_count_list, gc_dt, map_dt, sample_names, bin_size)
}
\arguments{
\item{feature_count_list}{Feature counts output list.}

\item{gc_dt}{GC content data table.}

\item{map_dt}{Mappability  data table.}

\item{sample_names}{Sample names.}

\item{bin_size}{Bin size.}
}
\value{
A data table with GC, mappability and count per bin.
}
\description{
Parse feature count output for bins and merge to gc and mappability.
}
