% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_bins.R
\name{parse_genespace_input}
\alias{parse_genespace_input}
\title{Parse genespace results directory to extract syntenic hits information.
Gets a list with one data table per pair of progenitors. Each data table contains all syntenic anchor pairs (syntenicHits.txt files from GENESPACE).}
\usage{
parse_genespace_input(genespace_dir)
}
\arguments{
\item{genespace_dir}{Path to a directory containing the syntenicHits output directory for GENESPACE.}
}
\value{
A list with one data table per pair of progenitors. Each data table contains all syntenic anchor pairs (syntenicHits.txt files from GENESPACE).
}
\description{
Parse genespace results directory to extract syntenic hits information.
Gets a list with one data table per pair of progenitors. Each data table contains all syntenic anchor pairs (syntenicHits.txt files from GENESPACE).
}
