% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_alignment}
\alias{plot_alignment}
\title{Plotting function for heal alignment}
\usage{
plot_alignment(
  heal_list,
  alignment,
  quick_view_sample = FALSE,
  output_dir = FALSE,
  n_threads = 1,
  add_bins = FALSE,
  prog_ploidy = 2,
  colour_map = FALSE,
  specific_chr = FALSE,
  return_list = FALSE
)
}
\arguments{
\item{heal_list}{Output list in heal format (such as output from count_heal_data()).}

\item{alignment}{A heal alignment object created with get_heal_alignment().}

\item{quick_view_sample}{The name of a sample to plot (as character)('FALSE' by default).}

\item{output_dir}{The name of a directory to write all plots to. Will create one if nonexistent.}

\item{n_threads}{Number of threads to use ('1' by default).}

\item{add_bins}{Add points for bins ('FALSE' by default). If "ref" the bins normalized copy number (divided by average (median by default)). If "alt" the bins overlapping with anchors are also added at the starting position of the reference anchor. If "FALSE" no bins are added.}

\item{prog_ploidy}{Ploidy of the progenitors (Assumed to be equal. '2' by default).}

\item{colour_map}{A vector of colours for each progenitor. If "FALSE" the colours are choosen using rainbow().}

\item{specific_chr}{A vector of characters indicating which chromosomes to plot (plots all by default).}

\item{return_list}{Logical: return a list of plots if quick_view_sample.}
}
\value{
Either nothing or a list of plots.
}
\description{
Plotting function for heal alignment
}
