% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_all_bins}
\alias{plot_all_bins}
\title{Plot counts and copy number for multiple samples together}
\usage{
plot_all_bins(
  heal_list,
  view_samples = "all",
  output_dir = FALSE,
  prog_ploidy = 2,
  plot_cn = FALSE,
  add_bins = TRUE,
  color_map = FALSE,
  method = "global",
  average = "median",
  average_list = FALSE,
  return_plot = FALSE,
  chr_label_size = 4,
  cn_label_size = 3,
  size_x_axis_title = 14,
  size_y_axis_title = 14,
  sample_name_size = 3,
  bin_point_alpha = 0.4,
  bin_point_size = 2,
  cn_line_width = 2,
  device = "pdf",
  ...
)
}
\arguments{
\item{heal_list}{List in heal format (such as output from count_heal_data()).}

\item{view_samples}{A vector of sample names to be plotted. Samples will be plotted in the order provided. Default value is "all", which shows all samples.}

\item{output_dir}{The name of a directory to write all plots to. Will create one if nonexistent.}

\item{prog_ploidy}{of the progenitors (Assumed to be equal. '2' by default).}

\item{plot_cn}{Logical: plot a line indicating infered copy number ('FALSE' by default in CN has been estimated).}

\item{add_bins}{Logical: plot counts for each bin ('TRUE' by default; normalized in plot_cn=TRUE).}

\item{color_map}{A vector of colors for each progenitor. If "FALSE" the colors are choosen using viridis().}

\item{method}{Which method was used to assign a copy number to each segment (if plot_cn=TRUE and add_bins!=FALSE). It should be the same method as the one used in get_copy_number() ('global', 'local' or 'manual'. 'global' by default).}

\item{average}{The method to compute average to normalize counts (if plot_cn=TRUE and add_bins!=FALSE). It should be the same method as the one used in get_copy_number() ('median' or 'mean'. 'median' by default).}

\item{average_list}{Same as in get_copy_number().}

\item{return_plot}{Logical: return the ggplot2 object ('FALSE' by default).}

\item{chr_label_size}{Size of the chromosome labels as input to ggplot2::geom_text size argument ('4' by default).}

\item{cn_label_size}{Size of the copy number labels as input to ggplot2::geom_text size argument ('3' by default).}

\item{size_x_axis_title}{Size of the x axis title as input to ggplot2::element_text size argument ('14' by default).}

\item{size_y_axis_title}{Size of the x axis title as input to ggplot2::element_text size argument ('14' by default).}

\item{sample_name_size}{Size of the sample names as input to ggplot2::geom_text size argument ('3' by default).}

\item{bin_point_alpha}{The transparency of the normalized bin count values as input to ggplot2::goem_point alpha argument ('0.4' by default).}

\item{bin_point_size}{Size of the normalized bin count values as input to ggplot2::goem_point size argument ('2' by default).}

\item{cn_line_width}{Thickness of the line showing the copy number as input to ggplot2::geom_path linewidth argument ('2' by default).}

\item{device}{Plot device (as argument to ggplot2::ggsave) ('pdf' by default).}

\item{...}{Any arguments you wish to pass to ggplot2::ggsave().}
}
\value{
Either nothing of a ggplot object showing several samples together. The plot is always printed.
}
\description{
Plot counts and copy number for multiple samples together
}
