% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_bins}
\alias{plot_bins}
\title{Plotting function for heal_list.}
\usage{
plot_bins(
  heal_list,
  view_sample = FALSE,
  output_dir = FALSE,
  n_threads = 1,
  prog_ploidy = 2,
  plot_cn = FALSE,
  add_bins = TRUE,
  color_map = FALSE,
  specific_chr = FALSE,
  return_list = FALSE,
  method = "global",
  average = "median",
  average_list = FALSE,
  linewidth = 2,
  ylim_max = 8,
  ...
)
}
\arguments{
\item{heal_list}{List in heal format (such as output from count_heal_data()).}

\item{view_sample}{The name of a sample to plot (as character)('FALSE' by default).}

\item{output_dir}{The name of a directory to write all plots to. Will create one if nonexistent.}

\item{n_threads}{Number of threads to use ('1' by default).}

\item{prog_ploidy}{Ploidy of the progenitors (Assumed to be equal. '2' by default).}

\item{plot_cn}{Logical: plot a line indicating infered copy number ('FALSE' by default in CN has been estimated).}

\item{add_bins}{Logical: plot counts for each bin ('TRUE' by default; normalized in plot_cn=TRUE).}

\item{color_map}{A vector of colors for each progenitor. If "FALSE" the colors are choosen using viridis().}

\item{specific_chr}{A vector of characters indicating which chromosomes to plot (plots all by default).}

\item{return_list}{Logical: return a list of plots if view_sample.}

\item{method}{Which method was used to assign a copy number to each segment (if plot_cn=TRUE and add_bins!=FALSE). It should be the same method as the one used in get_copy_number() ('global', 'local' or 'manual'. 'global' by default).}

\item{average}{The method to compute average to normalize counts (if plot_cn=TRUE and add_bins!=FALSE). It should be the same method as the one used in get_copy_number() ('median' or 'mean'. 'median' by default).}

\item{average_list}{Same as in get_copy_number().}

\item{ylim_max}{Maximum y axis value when plotting copy number.}

\item{...}{Any arguments you wish to pass to ggplot2::geom_point()}
}
\value{
Either nothing or a list of plots.
}
\description{
Plotting function for heal_list.
}
