% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_densities}
\alias{plot_densities}
\title{Plot empirical concordant densities}
\usage{
plot_densities(
  densities,
  quick_view_sample = FALSE,
  output_dir = FALSE,
  show_discordant = FALSE,
  heal_list = FALSE,
  alignment = FALSE,
  corrected_alignment = FALSE,
  ylim_max = FALSE,
  colour_vec = FALSE,
  prog_ploidy = 2,
  n_threads = 1
)
}
\arguments{
\item{densities}{A list of densities for each polyploid sample (output from get_concordant_density()).}

\item{quick_view_sample}{The name of a sample to plot (as character)('FALSE' by default).}

\item{output_dir}{The name of a directory to write all plots to. Will create one if nonexistent.}

\item{show_discordant}{Logical defining if bins involved in a discordant anchor set should be added on the plot.}

\item{heal_list}{Output list in heal format (such as output from count_heal_data()).}

\item{alignment}{A heal alignment object created with get_heal_alignment().}

\item{corrected_alignment}{A density corrected alignment ('FALSE' by default). If one is given and show_discordant==TRUE then the new state of corrected points is shown above the previous state.}

\item{ylim_max}{Maximum ylim value ('FALSE' by default). If FALSE then the value is set for each sample as 3 times the standard deviation above the mean.}

\item{colour_vec}{A vector of colours for each copy number class from 0 to the ploidy of the progenitors times the number of progenitors ('FALSE' by default). If "FALSE" the colours are choosen using rainbow().}

\item{prog_ploidy}{Ploidy of the progenitors (Assumed to be equal. '2' by default)}

\item{n_threads}{Number of threads to use ('1' by default).}
}
\value{
Nothing. Plots are shown and/or saved to output_dir.
}
\description{
Plot empirical concordant densities
}
