% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_heal_heat_map}
\alias{plot_heal_heat_map}
\title{Plot pairwise heatmap}
\usage{
plot_heal_heat_map(alignment, view_samples = FALSE, output_dir = FALSE)
}
\arguments{
\item{alignment}{A heal alignment object created with get_heal_alignment().}

\item{view_samples}{A vector of sample names to plot (as character)('FALSE' by default).}

\item{output_dir}{The name of a directory to write all plots to. Will create one if nonexistent.}
}
\value{
Nothing. Plots are shown and/or saved to output_dir.
}
\description{
Plot pairwise heatmap
}
