% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riparian_plots.R
\name{plot_riparian}
\alias{plot_riparian}
\title{Plot riparian style plots with copy number}
\usage{
plot_riparian(
  alignment,
  heal_list,
  genespace_dir,
  output_dir = FALSE,
  colour_scales = FALSE,
  theme = "light",
  legend_text_size = 5,
  legend_title_size = 7,
  legend_spacing = 2,
  title_size = 5,
  device_vector = c("pdf", "svg"),
  ...
)
}
\arguments{
\item{alignment}{A heal alignment object created with get_alignment().}

\item{heal_list}{List in heal format (such as output from count_heal_data()).}

\item{genespace_dir}{Path to a directory containing the syntenicRegion_coordinates.csv output file from GENESPACE.}

\item{output_dir}{The name of a directory to write all plots to. Will create one if nonexistent.}

\item{colour_scales}{A list with one entry per subgenome. Each entry must be named after a subgenome and contain a vector with the starting and ending colour value of the range. Defaults to red and green colour ranges.}

\item{theme}{Background settings. Options are 'dark' or 'light'. Default value is 'light'.}

\item{legend_text_size}{Size of the legend text as input to  ggplot2::element_text size argument ('5' by default).}

\item{legend_title_size}{Size of the legend title as input to  ggplot2::element_text size argument ('7' by default).}

\item{title_size}{Size of the title text as input to ggplot2::element_text size argument ('10' by default).}

\item{device_vector}{A vector with the file types you wish to have, as input ggplot2::ggsave device argument (c('pdf', 'svg') by default).}

\item{...}{Any arguments you wish to pass to ggplot2::ggsave().}
}
\value{
Nothing. Creates riparian style plots with copy number information.
}
\description{
Plot riparian style plots with copy number
}
