% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_functions.R
\name{read_aln_summary}
\alias{read_aln_summary}
\title{Read a heal alignment summary from a directory}
\usage{
read_aln_summary(aln_summary_dir, ...)
}
\arguments{
\item{aln_summary_dir}{A directory structured after a heal alignment summary (created by write_aln_summary()).}

\item{...}{Any argument you which to pass to fread (see data.table::fread documentation).}
}
\value{
A heal alignment summary i.e. A list with one data table per sample. Same as the output of summarize_aln().
}
\description{
Read a heal alignment summary from a directory
}
