% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_functions.R
\name{read_cn_summary}
\alias{read_cn_summary}
\title{Read a heal copy number summary from a directory}
\usage{
read_cn_summary(cn_summary_directory, ...)
}
\arguments{
\item{cn_summary_directory}{A directory structured after a CN summary (created by write_cn_summary()).}

\item{...}{Any argument you which to pass to fread (see data.table::fread documentation).}
}
\value{
A CN summary i.e. a list object with information about copy number for each sample. Same as the output of summarize_cn().
}
\description{
Read a heal copy number summary from a directory
}
