% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_functions.R
\name{read_heal_alignment}
\alias{read_heal_alignment}
\title{Read a heal alignment from a directory}
\usage{
read_heal_alignment(alignment_directory, ...)
}
\arguments{
\item{alignment_directory}{A directory structured after a heal alignment (created by write_heal_alignment()).}

\item{...}{Any argument you which to pass to fread (see data.table::fread documentation).}
}
\value{
A heal alignment i.e. A list with one data table per sample. Same as the output of get_heal_alignment().
}
\description{
Read a heal alignment from a directory
}
