% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_functions.R
\name{read_heal_list}
\alias{read_heal_list}
\title{Read a heal list from a directory}
\usage{
read_heal_list(heal_list_directory, ...)
}
\arguments{
\item{heal_list_directory}{A directory structured after a heal list (created by write_heal_list()).}

\item{...}{Any argument you which to pass to fread (see data.table::fread documentation).}
}
\value{
A heal list object i.e. a list with one element per progenitor containing at least a data table with counts in bins for each sample and GC and mappability for each bin.
}
\description{
Read a heal list from a directory
}
