% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_bins.R
\name{ref_alt_pair}
\alias{ref_alt_pair}
\title{Get the genome pairs of the synHits file.}
\usage{
ref_alt_pair(syn_hits_list, show_which_non_self = TRUE)
}
\arguments{
\item{syn_hits_list}{A list with the GENESPACE output files "synHits.txt.gz".}

\item{show_which_non_self}{Output a logical value being TRUE if the pair is not a self hit. Defaults is "TRUE".}
}
\value{
A list with one element per genome pair (including self). If show_which_non_self=TRUE the item is a logical value being TRUE if the pair is self hit, else it returns the base name of the file without suffix.
}
\description{
Get the genome pairs of the synHits file.
}
