% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_and_correct.R
\name{remove_short_spans}
\alias{remove_short_spans}
\title{Replace copy number spans shorter than user defined length by border value.}
\usage{
remove_short_spans(heal_list, max_length = 1, n_threads = 1)
}
\arguments{
\item{heal_list}{List in heal format (such as output from count_heal_data()).}

\item{max_length}{The maximal length at which a copy number span is to be replaced.}

\item{n_threads}{Number of threads to use. The function can process each chromosome in parallele, one subgenome at a time,
such that the maximal sensible value is that of the chromosome number in the subgenome with most chromosomes.}
}
\value{
A heal list with short CN spans replaced by neighbouring values if these are identical to each other.
}
\description{
Replace copy number spans shorter than user defined length by border value.
}
