% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riparian_plots.R
\name{sigmoid}
\alias{sigmoid}
\title{Function to  define a sigmoid style curve between two corrresponding block edges.}
\usage{
sigmoid(y, x_start, x_end, k = 7)
}
\arguments{
\item{x_start}{Start point on X.}

\item{x_end}{End point on X.}

\item{k}{k param of the sigmoidal function}

\item{y_vec}{Points to define along Y.}
}
\value{
A vector of x axis values to pair with the input y_vec to define the sigmoidal curve.
}
\description{
Function to  define a sigmoid style curve between two corrresponding block edges.
}
