% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize_aln}
\alias{summarize_aln}
\title{Summary of the alignment outcome for each polyploid sample}
\usage{
summarize_aln(alignment, n_threads)
}
\arguments{
\item{alignment}{A heal alignment object created with get_heal_alignment().}

\item{n_threads}{Number of threads to use ('1' by default).}
}
\value{
A list with one entry per polyploid sample. Each sample in turn contains a list with information about per progenitor and per chromosome.
The values are both a data table giving run length encoded style information along the chromosomes and another data table giving the proportion of the chromosomes
in each combination of copy number. There are also total distributions for each progenitor and for the sample overall.
}
\description{
Summary of the alignment outcome for each polyploid sample
}
