% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize_cn}
\alias{summarize_cn}
\title{Summary of the copy number analysis}
\usage{
summarize_cn(heal_list, n_threads = 1)
}
\arguments{
\item{heal_list}{A heal alignment object created with get_heal_alignment().}

\item{n_threads}{Number of threads to use ('1' by default).}
}
\value{
A list with one entry per sample. Each sample in turn contains a list with information about per progenitor and per chromosome.
The information at each chromosome is run length encoded information of copy number and the distribution of bins in each copy number.
There are also total distributions for each progenitor and for the sample overall.
}
\description{
Summary of the copy number analysis
}
