% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_functions.R
\name{write_aln_summary}
\alias{write_aln_summary}
\title{Save a heal alignment summary as a directory}
\usage{
write_aln_summary(alignment_summary, output_dir, ...)
}
\arguments{
\item{alignment_summary}{A list object created with summarize_aln()}

\item{output_dir}{The name of a directory to write the summary list to. Fails if the directory exists.}

\item{...}{Any argument you wish to pass to fwrite (see data.table::fwrite documentation). Writes as csv by default. /!\ Warning: Changing from default might cause issues in read_aln_summary() function.}
}
\value{
Nothing. Writes the data into a directory.
}
\description{
Save a heal alignment summary as a directory
}
