% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_functions.R
\name{write_heal_alignment}
\alias{write_heal_alignment}
\title{Save a heal alignment as a directory}
\usage{
write_heal_alignment(alignment, output_dir, ...)
}
\arguments{
\item{alignment}{A heal alignment object created with get_heal_alignment().}

\item{output_dir}{The name of a directory to write the alignment to. Fails if the directory exists.}

\item{...}{Any argument you wish to pass to fwrite (see data.table::fwrite documentation). Writes as csv by default. /!\ Warning: Changing from default might cause issues in read_alignment() function.}
}
\value{
Nothing. Writes the data into a directory.
}
\description{
Save a heal alignment as a directory
}
