% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_functions.R
\name{write_heal_list}
\alias{write_heal_list}
\title{Save a heal list as a directory}
\usage{
write_heal_list(heal_list, output_dir, ...)
}
\arguments{
\item{heal_list}{List in heal format (such as output from count_heal_data()).}

\item{output_dir}{The name of a directory to write the heal list to. Fails if the directory exists.}

\item{...}{Any argument you wish to pass to fwrite (see data.table::fwrite documentation). Writes as csv by default. /!\ Warning: Changing from default might cause issues in read_heal_list() function.}
}
\value{
Nothing. Writes the data into a directory.
}
\description{
Save a heal list as a directory
}
