% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_heal_data.R
\name{count_heal_data}
\alias{count_heal_data}
\title{Counts reads and combines results with GC and mappability information.}
\usage{
count_heal_data(
  input_dir,
  n_cores = 1,
  bin_size,
  paired_end,
  full_output = FALSE
)
}
\arguments{
\item{input_dir}{A healr input directory (see details).}

\item{n_cores}{Number of cores to use with featureCounts ('1' by default)}

\item{bin_size}{Bin size.}

\item{paired_end}{Logical: Is the data paired end.}

\item{full_output}{Logical: Do you want to also get the full featureCounts output ('FALSE' by default)}
}
\value{
A list with one element per progenitor containing at least a data table with counts in bins for each sample and GC and mappability for each bin.
}
\description{
Counts reads and combines results with GC and mappability information.
}
