% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_and_correct.R
\name{filter_bins}
\alias{filter_bins}
\title{Filter bins based on average per base mappability, read count and GC content.}
\usage{
filter_bins(
  heal_list,
  mappability_threshold = 0.9,
  gc_threshold = 2,
  count_threshold = 2,
  replace_by_NA = TRUE,
  log_file = FALSE
)
}
\arguments{
\item{heal_list}{List in heal format (such as output from count_heal_data()).}

\item{mappability_threshold}{Threshold average per bin mappability value below which bins are ignored ('0.9' by default).}

\item{gc_threshold}{How many standard deviations away from the GC content median to consider a bin as outlier. Set to 'FALSE' for no filtering ('2' by default).}

\item{count_threshold}{How many standard deviations above the count median to consider a bin as outlier. Set to 'FALSE' for no count filtering ('2' by default).}

\item{replace_by_NA}{Should outliers count values be set to NA ('TRUE' by default). Otherwise outliers are set to the threshold defined by count_threshold.}

\item{log_file}{A file to write the filtering statistics to ('FALSE' by default).}
}
\value{
A list with one filtered bins data table for each progenitor & the genes data tables if present (any full featureCounts outputs dropped).
}
\description{
Filter bins based on average per base mappability, read count and GC content.
}
