% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_copy_number.R
\name{get_copy_number}
\alias{get_copy_number}
\title{Infer copy number using Circular Binary Segmentation (from DNAcopy)}
\usage{
get_copy_number(
  heal_list,
  n_threads = 1,
  prog_ploidy = 2,
  method = "global",
  average = "median",
  average_list = FALSE,
  full_output = FALSE,
  manual_scale = FALSE
)
}
\arguments{
\item{heal_list}{Output list in heal format (such as output from count_heal_data() or filter_bins()).}

\item{n_threads}{Number of threads to use ('1' by default).}

\item{prog_ploidy}{Ploidy of the progenitors (Assumed to be equal. '2' by default).}

\item{method}{Which method to use to assign a copy number to each segment ('global', 'local' or 'manual'. 'global' by default). Global uses average over both subgenomes, local uses per subgenome average and manual uses values set by users in 'average_list'.}

\item{average}{Average measure used to assign a copy number in each segment ('median' or 'mean'. 'median' by default).}

\item{average_list}{A list with one element per sample with each containing one element per subgenome with values used to assign copy numbers.}

\item{full_output}{Logical: Do you want to also get the full DNAcopy output ('FALSE' by default).}

\item{manual_scale}{Manually set the value of multiples of average x ploidy to use during copy number assignment. The input should be a vector. The values in that vector will be used to set copy numbers going from 0 upwards.
For a vector of length N, copy numbers above N+1 will be assigned based on their multiple of the average. The function is set to 'FALSE' by default, meaning standard assignment.}
}
\value{
A list with one element per progenitor containing the bins and genes elements of the inpiut plus a data table with infered copy number per bin for each sample and GC and mappability for each bin.
}
\description{
Infer copy number using Circular Binary Segmentation (from DNAcopy)
}
