% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_bins}
\alias{plot_bins}
\title{Plotting function for heal_list.}
\usage{
plot_bins(
  heal_list,
  quick_view_sample = FALSE,
  output_dir = FALSE,
  n_cores = 1,
  prog_ploidy = 2,
  plot_cn = TRUE,
  add_bins = TRUE,
  colour_map = c("purple", "orange"),
  specific_chr = FALSE,
  return_list = FALSE
)
}
\arguments{
\item{heal_list}{Output list in heal format (such as output from count_heal_data())}

\item{quick_view_sample}{The name of a sample to plot (as character)('FALSE' by default).}

\item{output_dir}{The name of a directory to write all plots to. Will create one if nonexistent.}

\item{n_cores}{Number of cores to use ('1' by default).}

\item{prog_ploidy}{Ploidy of the progenitors (Assumed to be equal. '2' by default)}

\item{plot_cn}{Logical: plot a line indicating infered copy number ('TRUE' by default in CN has been estimated).}

\item{add_bins}{Logical: plot counts for each bin ('TRUE' by default; normalized in plot_cn=TRUE).}

\item{colour_map}{A vector of colours for each progenitor ('c(purple,orange)' by default).}

\item{specific_chr}{A vector of characters indicating which chromosomes to plot (plots all by default).}

\item{return_list}{Logical: return a list of plots if quick_view_sample.}
}
\value{
Either nothing or a list of plots.
}
\description{
Plotting function for heal_list.
}
